% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket_info.R
\name{create_bracket_url}
\alias{create_bracket_url}
\title{Create the URL for a particular year}
\usage{
create_bracket_url(year)
}
\arguments{
\item{year}{The year that you want to create a url for}
}
\description{
Create the URL for the year you want
}
