% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket_info.R
\name{get_bracket_data}
\alias{get_bracket_data}
\title{Get bracket data}
\usage{
get_bracket_data(year)
}
\arguments{
\item{year}{The year you want tournament data from.}
}
\description{
Pull Bracket data for NCAA Basketball Tournaments
}
