% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket_info.R
\name{get_region_names}
\alias{get_region_names}
\title{Get the names of the regions}
\usage{
get_region_names(html)
}
\arguments{
\item{url}{The url of the year you want.}
}
\description{
Get the names of the regions
}
